agent: collect the allowlists into allowlists.go
The agent runs as root with the docker socket mounted. Its trust boundary
was spread over 2215 lines of main.go: thirteen declarations deciding which
services, containers, images, repo directories and paths that process will
act on. Reading the whole boundary meant reading the whole file.

Pure move, same package. Each declaration keeps its doc comment verbatim;
allowedServices and allowedContainers additionally pick up the map alignment
gofmt wants, since the new file has to pass the linter with gofmt enabled.
No check changes, no renames, no reordering within a map. main.go loses 168
lines and gains none. No test file needed touching.

Splitting the rest of main.go is deliberately not part of this.
